requirements.txt
azure-common==1.1.26
azure-core==1.12.0
azure-storage-blob==12.8.0
azure-storage-common==2.1.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.1.1
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msrest==0.6.21
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
plotly==4.14.3
pyarrow==3.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
